Revisions to help AI prevent human victory
Changes were made in how the AI decides if it should make a preventative move over a move towards a victory state. At this stage, if the AI goes first - it should either always win or draw. (This is not fully tested so I will not call it perfect, but I have not been able to beat it as of this commit). The first move is hard coded to be at an optimal place and then future moves are decided by the same Breadth First Search technique as before.